chore(api): move Frameworks model to certification/shared
