+++ api/db/database-builder/factory/build-certification-course.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 
-import { Frameworks } from '../../../src/certification/configuration/domain/models/Frameworks.js';
 import { AlgorithmEngineVersion } from '../../../src/certification/shared/domain/models/AlgorithmEngineVersion.js';
+import { Frameworks } from '../../../src/certification/shared/domain/models/Frameworks.js';
 import { databaseBuffer } from '../database-buffer.js';
 import { buildSession } from './build-session.js';
 import { buildUser } from './build-user.js';
