+++ api/db/seeds/data/team-certification/cases/simple-pix-plus-droit-v2.js
@@ -1,6 +1,6 @@
 import dayjs from 'dayjs';
 
-import { Frameworks } from '../../../../../src/certification/configuration/domain/models/Frameworks.js';
+import { Frameworks } from '../../../../../src/certification/shared/domain/models/Frameworks.js';
 import { Assessment } from '../../../../../src/shared/domain/models/Assessment.js';
 import {
   PIX_DROIT_COMPLEMENTARY_CERTIFICATION_ID,
