+++ api/src/certification/enrolment/domain/models/Candidate.js
@@ -2,8 +2,8 @@
  * @typedef {import ('./Subscription.js').Subscription} Subscription
  */
 import { CertificationCandidatesError } from '../../../../shared/domain/errors.js';
-import { Frameworks } from '../../../configuration/domain/models/Frameworks.js';
 import { BILLING_MODES, SUBSCRIPTION_TYPES } from '../../../shared/domain/constants.js';
+import { Frameworks } from '../../../shared/domain/models/Frameworks.js';
 import { validate } from '../validators/candidate-validator.js';
 
 export class Candidate {
