+++ api/src/certification/session-management/domain/models/FinalizedSession.js
@@ -1,4 +1,4 @@
-import { Frameworks } from '../../../configuration/domain/models/Frameworks.js';
+import { Frameworks } from '../../../shared/domain/models/Frameworks.js';
 
 class FinalizedSession {
   constructor({
