+++ api/tests/acceptance/application/session/jury-certification-summaries-route_test.js
@@ -1,4 +1,4 @@
-import { Frameworks } from '../../../../src/certification/configuration/domain/models/Frameworks.js';
+import { Frameworks } from '../../../../src/certification/shared/domain/models/Frameworks.js';
 import {
   createServer,
   databaseBuilder,
