+++ api/tests/certification/configuration/integration/domain/usecases/find-certification-frameworks_test.js
@@ -1,5 +1,5 @@
-import { Frameworks } from '../../../../../../src/certification/configuration/domain/models/Frameworks.js';
 import { usecases } from '../../../../../../src/certification/configuration/domain/usecases/index.js';
+import { Frameworks } from '../../../../../../src/certification/shared/domain/models/Frameworks.js';
 import { SCOPES } from '../../../../../../src/certification/shared/domain/models/Scopes.js';
 import { databaseBuilder, expect } from '../../../../../test-helper.js';
