+++ api/tests/certification/enrolment/unit/infrastructure/candidates-import/CandidateData_test.js
@@ -1,5 +1,5 @@
-import { Frameworks } from '../../../../../../src/certification/configuration/domain/models/Frameworks.js';
 import { CandidateData } from '../../../../../../src/certification/enrolment/infrastructure/candidates-import/CandidateData.js';
+import { Frameworks } from '../../../../../../src/certification/shared/domain/models/Frameworks.js';
 import { expect } from '../../../../../test-helper.js';
 
 describe('Unit | infrastructure | candidates-import | CandidateData', function () {
