+++ api/tests/certification/results/integration/infrastructure/utils/csv/certification-results/get-division-certification-results-csv_test.js
@@ -1,7 +1,7 @@
 import dayjs from 'dayjs';
 
-import { Frameworks } from '../../../../../../../../src/certification/configuration/domain/models/Frameworks.js';
 import { getDivisionCertificationResultsCsv } from '../../../../../../../../src/certification/results/infrastructure/utils/csv/certification-results/get-division-certification-results-csv.js';
+import { Frameworks } from '../../../../../../../../src/certification/shared/domain/models/Frameworks.js';
 import { AutoJuryCommentKeys } from '../../../../../../../../src/certification/shared/domain/models/JuryComment.js';
 import { getI18n } from '../../../../../../../../src/shared/infrastructure/i18n/i18n.js';
 import { domainBuilder, expect } from '../../../../../../../test-helper.js';
