+++ api/tests/certification/results/integration/infrastructure/utils/csv/certification-results/get-session-certification-results-csv_test.js
@@ -1,6 +1,6 @@
-import { Frameworks } from '../../../../../../../../src/certification/configuration/domain/models/Frameworks.js';
 import { getSessionCertificationResultsCsv } from '../../../../../../../../src/certification/results/infrastructure/utils/csv/certification-results/get-session-certification-results-csv.js';
 import { AlgorithmEngineVersion } from '../../../../../../../../src/certification/shared/domain/models/AlgorithmEngineVersion.js';
+import { Frameworks } from '../../../../../../../../src/certification/shared/domain/models/Frameworks.js';
 import { AutoJuryCommentKeys } from '../../../../../../../../src/certification/shared/domain/models/JuryComment.js';
 import { getI18n } from '../../../../../../../../src/shared/infrastructure/i18n/i18n.js';
 import { domainBuilder, expect } from '../../../../../../../test-helper.js';
