+++ api/tests/tooling/domain-builder/factory/build-certification-result.js
@@ -1,6 +1,6 @@
-import { Frameworks } from '../../../../src/certification/configuration/domain/models/Frameworks.js';
 import { CertificationResult } from '../../../../src/certification/results/domain/models/CertificationResult.js';
 import { AlgorithmEngineVersion } from '../../../../src/certification/shared/domain/models/AlgorithmEngineVersion.js';
+import { Frameworks } from '../../../../src/certification/shared/domain/models/Frameworks.js';
 
 export function buildCertificationResult({
   id = 123,
