+++ api/tests/tooling/domain-builder/factory/certification/enrolment/build-candidate.js
@@ -1,5 +1,5 @@
-import { Frameworks } from '../../../../../../src/certification/configuration/domain/models/Frameworks.js';
 import { Candidate } from '../../../../../../src/certification/enrolment/domain/models/Candidate.js';
+import { Frameworks } from '../../../../../../src/certification/shared/domain/models/Frameworks.js';
 import { domainBuilder } from '../../../domain-builder.js';
 
 const buildCandidate = function ({
