+++ api/tests/tooling/domain-builder/factory/certification/results/build-global-mesh-level.js
@@ -1,5 +1,5 @@
-import { Frameworks } from '../../../../../../src/certification/configuration/domain/models/Frameworks.js';
 import { GlobalCertificationLevel } from '../../../../../../src/certification/results/domain/models/v3/GlobalCertificationLevel.js';
+import { Frameworks } from '../../../../../../src/certification/shared/domain/models/Frameworks.js';
 
 export const buildGlobalCertificationLevel = function ({
   reachedMeshIndex = 1,
